Extracts transformer into separate compilation unit
